\name{simulate}
\alias{simulate}
\title{Population genetics simulation}
\description{
This uses the shiny app to produce an interactive population genetics simulation.
}
\usage{
simulate(x)
}
\arguments{
  \item{x}{Text string indicating the simulation to run.  Currently the default and only option is "popgen"}
}
\details{
This simulation helps students understand the effects of selection, mutation, and genetic drift. The vignette included with the package is a \href{http://coleoguy.github.io/blackmon.pubs/popgen_sims.pdf}{lab activity guide}.  It is designed for a college level evolution course and should provide 1-2 hours of activities for students. 
}
\value{This function returns an interactive webpage.}
\references{
\url{http://coleoguy.github.io/}
}
\author{Heath Blackmon and Emma Goldberg}
\keyword{ pedagogical }
\keyword{ interactive }
\keyword{ teaching }
